rename: add new packages
