+++ src/auth/auth.service.ts
@@ -1,6 +1,6 @@
 import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { AuthCredentialDto } from './auth-credential.dto';
+import { AuthCredentialDto } from './dto/auth-credential.dto';
 import { UserRepository } from './user.repository';
 import * as bcrypt from "bcryptjs";
 import { JwtService } from '@nestjs/jwt';
